[AWIBOF-3856] merge rebuild performance related commits 	- buffer_pool implementation 	- add memory manager and buffer pool unit tests 	- apply buffer pool to rebuild read 	- apply cpu affinity changes for new buffer pool
Former-commit-id: 14e7dc0c145722da3a78abfa4117216cd672b394

+++ src/cpu_affinity/affinity_viewer.cpp
@@ -173,7 +173,7 @@ void
 AffinityViewer::Numa::_RegisterCpuRole(int cpuId, Role role)
 {
     int socketId = numa_node_of_cpu(cpuId);
-    if (socketId == INVALID_NUMA)
+    if (static_cast<uint32_t>(socketId) == INVALID_NUMA)
     {
         return;
     }
